+++ app/config-types.ts
@@ -1179,13 +1179,7 @@ export const isSegmentInConfig = (
 
 export const isColorInConfig = (
   chartConfig: ChartConfig
-): chartConfig is
-  | AreaConfig
-  | ColumnConfig
-  | LineConfig
-  | PieConfig
-  | ScatterPlotConfig
-  | BarConfig => {
+): chartConfig is Exclude<ChartConfig, { chartType: "table" | "map" }> => {
   return !isTableConfig(chartConfig) && !isMapConfig(chartConfig);
 };