+++ app/configurator/components/chart-options-selector.tsx
@@ -53,6 +53,7 @@ import {
   isMapConfig,
   isTableConfig,
   MapConfig,
+  RegularChartConfig,
   SortingType,
 } from "@/config-types";
 import {
@@ -561,7 +562,8 @@ const EncodingOptionsPanel = ({
         <ChartLayoutOptions
           encoding={encoding}
           component={component}
-          chartConfig={chartConfig}
+          // Combo charts use their own drawer.
+          chartConfig={chartConfig as RegularChartConfig}
           components={components}
           hasColorPalette={hasColorPalette}
           hasSubOptions={!!hasSubOptions}
@@ -658,7 +660,7 @@ const ChartLayoutOptions = ({
 }: {
   encoding: EncodingSpec;
   component: Component | undefined;
-  chartConfig: ChartConfig;
+  chartConfig: RegularChartConfig;
   components: Component[];
   hasColorPalette: boolean;
   hasSubOptions: boolean;